Source file
This file contains HTML, CSS, JAVA SCRIPT, JSON Files of browser extension to find dark patterns on different websites. It works on the algorithm of UI Guard which utilizes computer vision and natural language pattern whereas it also uses the machine learning algorithm which is in-built function that is Decision making, trees etc.